requirements.txt
python_dotenv>=1.0.1, <2.0.0
semantic-kernel>=1.0.0
openai>=1.0.0
azure-identity>=1.12.0
aiohttp>=3.8.0